Casting fix to eliminate warning messages at compile on 32-bit address machines Thanks to Piotr Luczek, the high priest of C.

+++ src/perfmon.c
@@ -3155,7 +3155,7 @@ static inline int process_smpl_entry(unsigned int native_pfm_index, int flags, p
     safety:
 #endif
     {
-      *pc = (caddr_t)(*ent)->ip;
+      *pc = (caddr_t)((size_t)((*ent)->ip));
       ++(*ent);
       return(0);
     }
